backend/pyproject.toml
[project]
name = "deepersystems-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "munraitoo13"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "blinker (==1.9.0)",
    "click (==8.1.8)",
    "dataclasses (==0.6)",
    "dnspython (==2.7.0)",
    "flask (==3.1.0)",
    "flask-cors (==5.0.1)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "markupsafe (==3.0.2)",
    "pymongo (==4.11.3)",
    "python-dotenv (==1.1.0)",
    "werkzeug (==3.1.3)",
    "pydantic (>=2.11.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
